+++ src/proxy/proxy.go
@@ -2,12 +2,15 @@ package proxy
 
 import (
 	"github.com/mosaicnetworks/babble/src/hashgraph"
+	"github.com/mosaicnetworks/babble/src/node/state"
 )
 
-// AppProxy ...
+// AppProxy defines the interface which is used by Babble to communicate with
+// the App
 type AppProxy interface {
 	SubmitCh() chan []byte
 	CommitBlock(block hashgraph.Block) (CommitResponse, error)
 	GetSnapshot(blockIndex int) ([]byte, error)
 	Restore(snapshot []byte) error
+	OnStateChanged(state.State) error
 }